requirements.txt
discord.py
aiohttp
pytesseract
python-dotenv
Pillow
google-cloud-vision
scrapfly-sdk 
